move CurrentMinimizerParameters to energy_optimization crate
