+++ crates/energy_optimization/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod current_minimizer;
+mod parameters;
+
+pub use parameters::CurrentMinimizerParameters;
