+++ crates/types/src/parameters.rs
@@ -5,7 +5,6 @@ use nalgebra::{Point2, Vector2, Vector3, Vector4};
 use serde::{Deserialize, Serialize};
 use serialize_hierarchy::SerializeHierarchy;
 
-use crate::joints::Joints;
 use crate::{
     joints::{arm::ArmJoints, head::HeadJoints, leg::LegJoints},
     kick_step::KickStep,
@@ -167,17 +166,6 @@ pub struct GameStateFilterParameters {
     pub tentative_finish_duration: Duration,
     pub distance_to_consider_ball_moved_in_kick_off: f32,
     pub whistle_acceptance_goal_distance: Vector2<f32>,
-}
-
-#[derive(Copy, Clone, Debug, Default, Deserialize, Serialize, SerializeHierarchy)]
-pub struct CurrentMinimizerParameters {
-    pub reset_threshold: f32,
-    pub reset_speed: f32,
-    pub reset_base_offset: f32,
-    pub optimization_speed: f32,
-    pub allowed_current: f32,
-    pub optimization_sign: Joints<f32>,
-    pub position_difference_threshold: f32,
 }
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize, SerializeHierarchy)]
